Refactor organism
