+++ web/src/components/organisms/Project/ProjectList/index.tsx
@@ -1,14 +1,5 @@
-import styled from "@emotion/styled";
-import { useParams } from "react-router-dom";
-
-import Button from "@reearth-cms/components/atoms/Button";
-import Content from "@reearth-cms/components/atoms/Content";
-import Icon from "@reearth-cms/components/atoms/Icon";
-import Input from "@reearth-cms/components/atoms/Input";
 import ProjectCreationModal from "@reearth-cms/components/molecules/Common/ProjectCreationModal";
-import Greeting from "@reearth-cms/components/molecules/Dashboard/Greeting";
-import MoleculesProjectList from "@reearth-cms/components/molecules/Dashboard/MoleculesProjectList";
-import { useT } from "@reearth-cms/i18n";
+import Dashboard from "@reearth-cms/components/molecules/Dashboard";
 
 import useHooks from "./hooks";
 
@@ -17,10 +8,6 @@ export interface Props {
 }
 
 const ProjectList: React.FC<Props> = ({ handleWorkspaceModalOpen }) => {
-  const t = useT();
-  const { Search } = Input;
-  const { workspaceId } = useParams();
-
   const {
     handleProjectSearch,
     handleProjectCreate,
@@ -32,29 +19,12 @@ const ProjectList: React.FC<Props> = ({ handleWorkspaceModalOpen }) => {
 
   return (
     <>
-      <PaddedContent>
-        <Greeting />
-        <ActionHeader>
-          <Search
-            onSearch={handleProjectSearch}
-            placeholder={t("input search text")}
-            allowClear
-            type="text"
-            style={{ width: 264 }}
-          />
-          <ButtonWrapper>
-            <Button onClick={handleWorkspaceModalOpen}>{t("Create a Workspace")}</Button>
-            <Button onClick={handleProjectModalOpen} type="primary" icon={<Icon icon="plus" />}>
-              {t("New Project")}
-            </Button>
-          </ButtonWrapper>
-        </ActionHeader>
-        <MoleculesProjectList
-          projects={projects}
-          workspaceId={workspaceId}
-          handleProjectModalOpen={handleProjectModalOpen}
-        />
-      </PaddedContent>
+      <Dashboard
+        handleProjectSearch={handleProjectSearch}
+        handleProjectModalOpen={handleProjectModalOpen}
+        handleWorkspaceModalOpen={handleWorkspaceModalOpen}
+        projects={projects}
+      />
       <ProjectCreationModal
         open={projectModalShown}
         onClose={handleProjectModalClose}
@@ -63,25 +33,5 @@ const ProjectList: React.FC<Props> = ({ handleWorkspaceModalOpen }) => {
     </>
   );
 };
-
-const PaddedContent = styled(Content)`
-  margin: 16px;
-  background-color: #fff;
-  min-height: 100%;
-`;
-
-const ActionHeader = styled(Content)`
-  max-width: 1200px;
-  margin: auto;
-  padding: 16px;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const ButtonWrapper = styled.div`
-  Button + Button {
-    margin-left: 8px;
-  }
-`;
 
 export default ProjectList;
